add link connection WO
