+++ menghua_co/hooks.py
@@ -250,8 +250,7 @@ doctype_js = {
 
 doc_events = {
     "Work Order": {
-        "after_insert": "menghua_co.custom.work_order.make_manufacturing_order_auto",
-        "on_update": "menghua_co.custom.work_order_dashboard.after_work_order_creation"
+        "after_insert": "menghua_co.custom.work_order.after_insert_combined"
     },
     "Sales Order": {
         "after_insert": "menghua_co.custom.sales_order_dashboard.after_sales_order_creation"
